Fix compiler warnings.

+++ examples/utils/snitch.cpp
@@ -383,7 +383,7 @@ void compress(State& state, const std::string& folder, const std::string& archiv
         // write the file in multiple parts
         for (s64 offset = 0; offset < file.size(); offset += large_block_size)
         {
-            u64 size = std::min(large_block_size, file.size() - offset);
+            u64 size = std::min(large_block_size, u64(file.size() - offset));
 #ifdef DISABLE_MMAP
             file.read(buffer.data(), size);
             output.write(buffer.data(), size);
